Default settings to use the package as a node dependency
