+++ .editorconfig
@@ -1,0 +1,22 @@
+# EditorConfig is awesome: http://EditorConfig.org
+
+root = true
+
+# General Options
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+# JavaScript
+[*.js]
+indent_size = 2
+indent_style = space
+
+# Markdown
+[*.md]
+indent_size = 4
+trim_trailing_whitespace = false
