+++ .gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+npm-debug.log
+node_modules
