+++ package.json
@@ -1,0 +1,35 @@
+{
+  "name": "angular-facebook",
+  "version": "1.0.0",
+  "description": "An AngularJS module to handle Facebook API",
+  "author": "Daniel Acuña <@moustacheful>",
+  "contributors": [
+    {
+      "name": "Franco Moya <@iamravenous>"
+    }
+  ],
+  "main": "dist/angular-facebook.min.js",
+  "license": "MIT",
+  "homepage": "https://github.com/iamravenous/angular-facebook#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/iamravenous/angular-facebook.git"
+  },
+  "bugs": {
+    "url": "https://github.com/iamravenous/angular-facebook/issues"
+  },
+  "keywords": [
+    "angularjs",
+    "angular",
+    "facebook",
+    "sdk"
+  ],
+  "devDependencies": {
+    "del": "^2.2.0",
+    "gulp": "^3.9.1",
+    "gulp-load-plugins": "^1.2.2",
+    "gulp-rename": "^1.2.2",
+    "gulp-uglify": "^1.5.3",
+    "run-sequence": "^1.1.5"
+  }
+}
